+++ fortscale/fortscale-webapp/src/main/java/fortscale/web/rest/entities/activity/OrganizationActivityData.java
@@ -5,9 +5,13 @@ package fortscale.web.rest.entities.activity;
  * 22/05/2016
  */
 public class OrganizationActivityData {
+
     public static class LocationEntry extends BaseLocationEntry {
-        public LocationEntry(String country, Integer count) {
+
+        public LocationEntry(String country, Double count) {
             super(country, count);
         }
+
     }
+
 }
